amazon-web-scraper/src/db.py
import os
from dotenv import load_dotenv
import psycopg
from psycopg.rows import dict_row

load_dotenv()

class Database:
    def __init__(self):
        self._dsn = (
            f"host={os.getenv('POSTGRES_HOST')} "
            f"port={os.getenv('POSTGRES_PORT')} "
            f"dbname={os.getenv('POSTGRES_DB')} "
            f"user={os.getenv('POSTGRES_USER')} "
            f"password={os.getenv('POSTGRES_PASSWORD')}"
        )

    def connection(self):
        print('Connecting to database...')
        return psycopg.connect(self._dsn, row_factory=dict_row)