amazon-web-scraper/pyproject.toml
[project]
name = "amazon-web-scraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "psycopg[binary]>=3.2.0",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "pydantic>=2.12.5",
    "langchain-openai>=1.1.10",
]
